Fix linking of thirdparty_sleuthkit
The thirdparty_sleuthkit_fs_cpp object library actually depends
on thirdparty_sleuthkit_auto but we are not explicitly specifying
that depedency.
Sometimes it works, sometimes it doesn't (like on oss-fuzz).

Add the correct dependency and also change thirdparty_sleuthkit_fs_cpp
to be a STATIC library, because linking to thirdparty_sleuthkit_auto
causes a circular dependency (which is also present upstream),
and CMake can only deal with that if every library involved is STATIC.