Do not run clang-tidy on third party libraries
We don't want to run clang-tidy on third party libraries,
so we set that variable after having declared the third party library
targets. This skips setting the relative target property,
and CMake will build normally.

Setting the CMAKE_CXX_CLANG_TIDY variable in a global scope
is not needed, since all the targets we are interested in
are created in the topmost main() function scope.

Finally, only set the CXX variant of the variable,
since the osquery code doesn't contain C.