Update the ATC table `path` column check to be case insensitive
In sqlite, column names are case insensitive. But, in #6278 I added logic to _add_ the `path` column to ATC configs. Unfortunately, this logic was not case sensitive, and sometimes try to create a table with sql akin to:

```sql
CREATE TABLE atc_downcase(`path` TEXT, `Anum` TEXT, `Path` TEXT)
```

The double `path` and `Path` are a fatal sql error

Refine the `path` checks to be case insensitive.

Fixes #7441