Fix and improve Audit message header parsing
The logic parsing the Audit message header was a bit too lenient
towards a malformed prefix and so it was possible to cause a crash.
Improve the checks done to avoid crashes.

Also improve how the message "subtype" (seccomp, AppArmor, SELinux)
is detected, avoiding multiple copies of the Audit message string
and avoiding multiple scans.