Add third party libraries target
Add a helper target which can be used to compile all the third party
libraries.

This can be useful if you need to run a tool that instruments the build,
but you're only interested on the osquery portion of the code.

An example for this is the creation of a database for CodeQL.