Fix user_time and system_time unit in processes table on M1
Use the time base conversion coming from mach_timebase_info
also for the cpu time values coming from the libproc API.

The user_time and system_time columns values are taken
from the proc_pid_rusage API values of ri_user_time and
ri_system_time.
The unit of those two values was incorrectly considered
to always be nanoseconds, but they are mach ticks,
much like the start_time that is later calculated
via mach_absolute_time.

While on x86_64 the time base info used to convert ticks to nanoseconds
returns a numerator and denominator that are both 1,
on M1 they are not (125 and 3 is one example).