Use additional instead of index for admindir in deb_packages
Although both index and additional add the column to the primary key,
index is meant to mark columns that make the row unique.
So the column values should contain unique values.
additional instead marks a column that when provided in a where clause
changes the table behvior, which is more correct in this case,
since the packages will be parsed from a different path than the default.