Fix crash due to interaction between distributed and config plugin (#7504)
Move the distributed plugin RocksDB metadata to a different domain
which is not used or touched by the config plugin.

Also do not acquire the pending queries multiple times just to count
them. Acquire them once, use the obtained vector to count them
and then use them to proceed executing the queries.