Warn about setting CLI_FLAGs in the config
CLI_FLAG should not be set/updated in a config file,
but set once via a flagfile or passed at the process startup.
In the next release the ability to do so will be completely removed.

Also change the disable_tables and enable_tables flags
to CLI_FLAG since changes to them won't be reflected at runtime,
but it's also not desirable to have them change at runtime
for security reasons.