Fix thrift server shutting down when dropping privileges
This is a regression from last time we updated Thrift
and seems to affect only Linux.

Thrift had a patch to ignore the EINTR error
when waiting for connections.
This is because on Linux when dropping privileges
we call libc APIs to change user and groups
that require all threads to change credentials too.
The mechanism uses signals,
which interrupt other syscalls (like a poll) with EINTR.

Thrift has some limited retries for the EINTR case,
but it's pretty easy to hit.

Instead of patching Thrift it's possible to call
the underlying syscalls to change user and group
for a single thread, because the signaling mechanism
is implemented by the libc library only.
This was already partially done but missing
on a couple more function calls.