Fix SchedulerTests.test_scheduler_drift_accumulation flakyness
The test was not resetting the shutdown status,
which gets set everytime the scheduler does all the steps
it was requested to do.
Due to a recent change the scheduler was correctly
not executing queries anymore if the shutdown was requested.

This often caused very low execution time
and the test expecting at least 1ms drift was failing.