Initialize users and groups services on all tests that need them
Some of the tests making queries on tables that were
indirectly making use of the users and groups cached information
were sometimes failing because the needed services were not started.
This starts the appropriate services and makes sure to reset
the futures needed to access the cache, so that all the tests
that do not initialize the services on their own will fail,
instead of relying on a previous initialization.