Do not catch table or registry exceptions when running tests
Change the table_exceptions and registry_exceptions flags
to ignore_table_exceptions and ignore_registry_exceptions,
also changing their default to false.
This way all the tests on tables will fail if an exception
is thrown during the querying of a table.

They will be set to true though, when the Initializer
is used to start an osquery or extension instance.