Revert "Fixes osquery tls connections to aws kinesis when tls_server_certs is set (#7450)"
This reverts commit 97815bd68c36df4c997dbeff3a9c113f5a8d24ba.

This doesn't actually fix the issue the PR was set to fix.
The certificates are always needed to be set,
the difference though is that it might be desirable to have
a set of configurations for internal use,
to have osquery connect to a fleet manager,
and another set for other services, like AWS, Azure etc.