test: Fix Mdfind.test_sanity flakyness
Instead of searching for a hardcoded file name,
which might not be indexed yet, sample what is indexed,
then choose one of those to further test the API.

Also remove the Spotlight running test since it's incorrect.
The process that's doing the indexing is mdworker,
but we cannot check if it's actively indexing or not.
Assume actively indexing is the correct state and so
we shouldn't accept an empty result for the initial wide query.