Port `memory_devices` table to Windows (#7633)
* Initial port of memory devices table to Windows

* Fix size column in memory_devices table

* Add sanity test for memory_devices

* Make memory_devices sanity test root only

* Change units for memory device size

* Fix integration test cmake list

* Implement memory_devices table for M1 Macs

As M1 Macs are SoC's with soldered memory, most columns are not
available or don't make sense. This this includes the type and size
columns as the two that makes sense to include.

* Implement memory_devices table for M1 Macs

As M1 Macs are SoC's with soldered memory, most columns are not
available or don't make sense. This this includes the type and size
columns as the two that makes sense to include.

* Fix library updates