Fix process_file_events subscriber being incorrectly initialized
The process_file_events subscriber was not finding its data
in the database, so no event was shown.

The setUp() function shouldn't be overridden
by the implementation of the subscriber,
otherwise vital information about the subscriber
and where it should search its data in the database is not setup.
Only init() should be used.