Relax osquery safe permissions check
Ignore TrustedInstaller and CREATOR OWNER ACLs on both
the osqueryd binary and the parent folder.

TrustedInstaller is a system account, its ACL normally added by installers.
It's also present on the Windows folder,
owning it and having full access, so consider it safe.

CREATOR OWNER is placeholder ACL, it doesn't actually
give any permission to the owner.
When a new file or folder is created in the folder that contains
that ACL, then an additional ACL will be applied,
that has as a user the owner of the parent folder,
and as permissions the permissions of the CREATOR OWNER ACL.
Therefore consider it safe.

Co-Authored-By: Joe Sweeney <joe.sweeney@trailofbits.com>
Co-Authored-By: Stefano Bonicatti <stefano.bonicatti@gmail.com>