Fix deadlock when logging happens during a database reset
A deadlock could happen if a log relay thread
was trying to serialize logs into the database
when a database reset was being attempted.

Since the log relay thread is started by the same thread
that executes the database reset, the scheduler thread,
ensure that the log relay thread has finished its work
before doing a database reset on the next scheduler loop.

Also ensure that when the scheduler is finishing its work,
to permit osquery to exit, we wait on the log relaying thread
if it's still running to prevent race conditions
and possible crashes on shutdown.

Finally remove the relayStatusLog call from the watcher process,
it's a no-op since there's no logger plugin active.