Report empty snapshot at new epoch (#7803)
This change adds logging of an empty result set in batch format for an initial run (or new epoch) if there are no results. Previously the no results case was ignored and not logged even on an initial run / new epoch.

This makes the initial run / new epoch for differential queries more consistent with the snapshot query behavior, and it ensures there is always a "counter=0" log entry for each epoch in batch format. It makes it possible for users to differentiate between a query not running and the query not producing any results.

More context is in https://github.com/osquery/osquery/issues/7799

### Backward Compatibility
This should not lead to significantly more log events than before, as it is adding at most one log per epoch change.

There's a risk that implementations won't handle having both "added" and "removed" lists be empty correctly, but this risk seems low enough as it's already valid/expected for either of these lists to be empty independently. Osquery documentation never promises not to emit an empty entry like this, and its meaning seems reasonably expected/clear.