Fix username field for managed_policy table (#7944)
Currently the username field in managed_policy table is always set to "." for a user policy. This changes the field to instead be the actual username (or empty if it is a system policy), which is the expected behavior. 

The issue is that directory paths from `listDirectoriesInDirectory` end in "/"  (see [here](https://github.com/osquery/osquery/blob/cb65b6e9d6b2ca15278da8782fcad8507a16246d/osquery/filesystem/filesystem.cpp#L364)), which causes their filename and stem to be "." due to the Boost behavior described [here](https://www.boost.org/doc/libs/1_68_0/libs/filesystem/doc/reference.html#path-filename)

Tested manually with a managed machine using osqueryi. Also uncommented the table test, though it doesn't actually test anything on a non-managed machine (ie in CI) because the plist files are not present.