watchdog: Use virtual cores to calculate CPU utilization limit
Before this the CPU utilization limit, when on a physical host
with virtual cores, was actually half of what was configured
through the watchdog_utilization_limit flag,
because the code was counting only the physical cores,
which normally are half of the virtual ones.

On VMs this could've been worse; now we use all the cores
to calculate the limit.