fix: Correct the aws_sts_region behavior
The `aws_sts_region` flag was controlling the region used
by services using STS, like Kinesis and Firehose,
instead of controlling which region to generate the STS credential from.
This changes the behavior of the flag so it controls the STS credentials
region, and it also adds two new flags, `aws_kinesis_region` and
`aws_firehose_region` to control the relative services regions.