Added cloud_id to ycloud_instance_metadata - the vm metadata table for Yandex Cloud (#8086)
Added `cloud_id` column for `ycloud_instance_metadata` table. 

`cloud_id` is a container for resources in Yandex Cloud which is similar to a "project" in Google Cloud Platform. This change will allow people that use osquery in Yandex Cloud to get more information about the resource path of their virtual machines.