ci: Correct job order
The check_source_code job was waiting for the same job 2 times,
while it should've also waited for the code style job.

Also add a dependency in the build jobs to the libraries manifest check.
The general idea is to have quick checks go first and fail everything
immediately if something is wrong.