cmake: Pass the osquery python path to googletest
Although googletest uses python only for its own tests,
it still searches for it and the search cannot be removed.
In some configuration, like when using pyenv, CMake requires a shell to be able to
find Python; this might not be present if run via the VsCode CMake
extension.
While we can pass the Python3_EXECUTABLE path to osquery, googletest
will still attempt to search Python in the system, slowing down the
configuration phase.