Implement `battery` table for Windows (#8267)
Almost all of the columns are able to line up with those already available on macOS.

This has been tested on a single Windows 11 Pro laptop. All columns look correct except for `cycle_count` in which this device reports `0`. We seem to be using the correct API though and will hopefully get values on other devices.

Testing also performed on a macOS laptop with Windows 10/11 in VMs - the table successfully reports the VMWare virtual battery when that is configured and no results (along with appropriate logging) when there is no virtual battery.