Add --yara_sigurl_authenticate flag
This allows osquery users to host their yara rules behind an authenticated endpoint.

Allow authentication for yara sigurl requests by enabling this flag. When enabled, osquery will send a POST request with a body containing the node key as in other osquery TLS requests. Server response is expected to be the same as unauthenticated requests: a plaintext yara signature file.