Add support for msix packages (#8585)
https://github.com/fleetdm/fleet/issues/27199

This code scans the new registry location for MSIX packages and adds them to the `programs` osquery table. MSIX is the new an improved format designed to replace older formats like MSI, AppX, EXE installers. The Windows App Store uses it.  To learn more about it https://learn.microsoft.com/en-us/windows/msix/overview

The important bits to know about them is that they are scoped by user (see comment below). And a lot of the meta data around the application is contained in a file called `appxmanifest.xml`.

This code walks through the registry and grabs the installed applications. It also parses bits of the appxmanifest.xml file to get relevant data out.