Added UpgradeCode to programs table (#8587)
For issue https://github.com/fleetdm/fleet/issues/27759
Related to PR https://github.com/osquery/osquery/pull/8585

There are a few identifiers for windows programs. `UpgradeCode` is the unique identifier for MSI applications that is stable between version upgrades. It is different than the `identifying_number` column (which is typically `ProductCode`), as this changes between versions.

https://learn.microsoft.com/en-us/windows/win32/msi/productcode
https://learn.microsoft.com/en-us/windows/win32/msi/upgradecode