Filter the Win32_Processor query to only required fields (#8598)
Querying Win32_Processor for all fields has a performance penalty of 1.1s per CPU core - this is mostly unnoticable on small boxes, but on larger boxes can actualloy have a significant impact on startup performance.  Even on an 8 core system, this takes the startup time for osqueryi from 8.5s to 0.2s.