Add detection for ARM CPUs when running in x86 emulation (#8572)
The installer for osquery ships an x86 executable that runs on ARM because of the emulation layer built-in to Windows. Unfortunately because it's running under emulation, Windows will report a fake `x86_64` CPU.

This will make sure we check for WOW64 emulation information, which will return the actual CPU type.