Revert "Revert "Update Windows runner version in hosted_runners.yml (#8618)"" (#8636)
Reverts osquery/osquery#8633

Originally we needed to do this to get 5.18.1 working on Windows ARM. Now we are bringing back the newer runners so that Windows unit and integration tests will run successfully.

We will still need to figure out why those Windows builds are crashing on ARM with this runner.