Fix `NSInvalidArgumentException` when querying `connected_displays` (#8628)
This PR resolves an exception when querying `connected_displays` if one or more displays are rotated.

The content of `spdisplays_rotation` is `"spdisplays_supported"` when rotation is supported but the display is not rotated (i.e., 0 degree rotation) and either `90`, `180`, or `270` when the display is rotated. An example of both cases is available [here](https://github.com/osquery/osquery/issues/7486#issuecomment-1050680675).