Add new darwin certificate trust settings table (#8715)
This PR adds a new table for darwin devices, the
`certificate_trust_settings` table to acquire the configured cert trust
settings of each certificate domain for increased visibility into the
status of installed certificates. This data is similar to executing
`security dump-trust-settings -d` (for the 'admin' domain), or `security
dump-trust-settings -s` (for the 'system' domain).

Something that `security` fails to do, is correctly report the
configured trust settings of MDM distributed certificates. At least in
some cases, it will report that 0 settings are configured regardless of
which settings are set up.

For this table I'm implementing [Apple's Security
framework](https://developer.apple.com/documentation/security/sectrustsettingscopycertificates(_:_:)?language=objc)
to copy the certificate trust settings for a cert domain, and enumerate
the returned data.

When I was building the table, I could see that the array returned from
`SecTrustSettingsCopyTrustSettings` included the trust setting policy
name: `kSecTrustSettingsPolicyName`, but when building the table it
would fail for the use of an undeclared identifier. This is why I am
manually defining this key identifier.

Here is some redacted output:
```
osquery> SELECT * FROM certificate_trust_settings WHERE trust_domain = 'admin';
+------------------------------------------------------------------+------------------------------------+--------------+-------------------+-------------------+---------------------+-----------------+--------------+
| common_name                                                      | serial_number                      | trust_domain | trust_policy_name | trust_policy_data | trust_allowed_error | trust_key_usage | trust_result |
+------------------------------------------------------------------+------------------------------------+--------------+-------------------+-------------------+---------------------+-----------------+--------------+
| Charles Proxy CA (13 Sep 2024, )        | 000000000000                       | admin        | sslServer         |                   | -2147408896         |                 | trusted_root |
| Zscaler Root CA                                                  | 000000000000000000                 | admin        | sslServer         |                   | -2147408896         |                 | trusted_root |
+------------------------------------------------------------------+------------------------------------+--------------+-------------------+-------------------+---------------------+-----------------+--------------+
```