Fix crash when querying carves table with carves larger than 2GB (#8732)
Related issue: Resolves https://github.com/fleetdm/fleet/issues/34610

Fixing a bug that was discovered by upload carves >2gb.
It appears that the check for `IsInt` was not sufficient to check for
larger integers. byte values of 2gb+ are greater than the max of int32.
Switched to int64 to handle this properly.

---------

Co-authored-by: Zach Wasserman <zach@fleetdm.com>