Quit carving when sending a block fails (#8733)
Currently we log (only in --verbose) an error and then continue sending
blocks. But there's not really much point in continuing to send blocks
if the server failed a block because it won't be able to piece the
archive back together. What we really need is a refactor to allow proper
retries, but this at least makes the current behavior more clear and
removes inefficient continued block sends that won't lead to a
successful carve.