tables: Fix vscode_extensions visibility for LDAP/SSSD users
The usersFromContext helper is only capable of looking up users with a
single constraint (uid).  However, in order to return remote users on
linux, include_remote=1 must be specified.  This made it so that anything
leveraging usersFromContext to get a list of users, was not able to
list remote users.

see: https://github.com/osquery/osquery/issues/8707

This PR updates the following
- makes genUsers accessible via a new header, users.h
- updates vscode_extensions table to use the genUsers
  function instead of usersFromContext.
- updates the vscode_extensions table to include a hidden column
  include_remote, mirroring the one in the users table, this allows
  us to pass the query context as is when listing users

Signed-off-by: Brian McKinney <brian.mckinney@elastic.co>